Undefine major and minor through <sys/sysmacros.h>
<sys/sysmacros.h> is included through some other headers.
This results in major(x) being resolved to gnu_dev_major(x)
which is an expression in a constructor initializer list.
This patch includes <sys/sysmacros.h> at the very beginning,
and undefines major and minor to avoid them to be resolved
to gnu_dev_major(x) or gnu_dev_minor(x).